docker/requirements.txt
boto3==1.20.5
PyGithub==1.54.1
aurora-data-api==0.4.0
requests==2.28.0